First Roop walnut UAV data analysis
This work is done before, but it's the first time to be uploaded to github for version control. 
This includes both data preprocessing and model building. 
Current highest R^2 of regression model is 0.7824 from ridge regression model, and highest accuracy for 4-class classification model is 73.62%, and the highest accuracy for 3-class classification model is 82.00% from random forest classification.