Model building based on Sentinel-2 data
This is the second step after preprocessing Sentinel-2 data. 
This work has been done before, but it's the first time to be uploaded to github for version control. 
This work is based on the sentinel images resampled to 10m resolution. This is based only on the 2017 Roop walnut site data. 
The highest R^2 for the regression model is 0.47, the highest accuracy for 4-class classification is 71.5%, and the highest accuracy for 3-class classification is 76.5%.