Roop 2017 Sentinel process with match dates 20m
This is based on previous codes, but the used sentinel data is in 20m resolution, while previous work used sentinel data in 10m resolution. 
This is the processing work, and the only output is a data frame containing 5 dates' reflectance, vegetation index, swp, and weather data. The output will be used as the input of other model building code.